requirements.txt
accelerate==1.3.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
attrs==25.1.0
bitsandbytes==0.42.0
certifi==2025.1.31
charset-normalizer==3.4.1
datasets==3.3.0
diffusers==0.32.2
dill==0.3.8
filelock==3.17.0
frozenlist==1.5.0
fsspec==2024.12.0
huggingface-hub==0.28.1
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.5
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
numpy==2.2.3
packaging==24.2
pandas==2.2.3
peft==0.14.0
pillow==11.1.0
propcache==0.2.1
psutil==7.0.0
pyarrow==19.0.0
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scipy==1.15.1
setuptools==75.8.0
six==1.17.0
sympy==1.13.1
tokenizers==0.21.0
torch==2.7.0.dev20250215
torchaudio==2.6.0.dev20250215
torchvision==0.22.0.dev20250215
tqdm==4.67.1
transformers==4.48.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
